requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
logging==0.4.9.6
MarkupSafe==2.1.5
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.2
